No errors building version.

+++ lib/main/sensors.hpp
@@ -18,7 +18,7 @@ private:
 public:
     Quaternion q;                // [w, x, y, z]         quaternion container
     VectorFloat gravity;         // [x, y, z]            gravity vector
-    double ypr[3];               // [yaw, pitch, roll]   yaw/pitch/roll container
+    float ypr[3];               // [yaw, pitch, roll]   yaw/pitch/roll container
 
     esp_err_t initialize();
     void read_data();
@@ -29,9 +29,9 @@ class bmp280_handler_t {
     bmp280_t bmp_dev;
 
 public:
-    double pressure;
-    double temperature; 
-    double humidity;
+    float pressure;
+    float temperature;
+    float humidity;
 
     esp_err_t initialize();
     void read_data();
